invio 12
aggiunte dependencies bootStrap + elencocomunibootstrap.html +
nuovi packages web con risorse utili ad interfacciarsi col modulo
bootstrap per la creazione di pagine dinamiche Thymeleaf

+++ src/main/java/it/epicode/be/energy/security/controller/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
 
 	@GetMapping("/user/{id}")
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public ResponseEntity<User> findById(@PathVariable(required = true) Integer id) {
+	public ResponseEntity<User> findById(@PathVariable(required = true) Long id) {
 		Optional<User> find = userService.findById(id);
 
 		if (find.isPresent()) {
