+++ src/main/java/it/epicode/be/energy/security/service/UserService.java
@@ -14,7 +14,7 @@ public class UserService {
 	@Autowired
 	private UserRepository userRepository;
 
-	public Optional<User> findById(Integer id) {
+	public Optional<User> findById(Long id) {
 		return userRepository.findById(id);
 	}
